[MPLUGIN-432] Improve site build
- rename duplicate packages - cause NPE in javadoc on clean project
- remove reports defined in parent pom
- aggregate reports should be not inherited
