+++ maven-plugin-plugin/src/main/java/org/apache/maven/plugins/plugin/descriptor_old/EnhancedPluginDescriptorBuilder.java
@@ -1,4 +1,4 @@
-package org.apache.maven.plugins.plugin.descriptor;
+package org.apache.maven.plugins.plugin.descriptor_old;
 
 /*
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -25,6 +25,7 @@ import org.apache.maven.plugin.descriptor.MojoDescriptor;
 import org.apache.maven.plugin.descriptor.Parameter;
 import org.apache.maven.plugin.descriptor.PluginDescriptor;
 import org.apache.maven.plugin.descriptor.PluginDescriptorBuilder;
+import org.apache.maven.plugin.plugin.report_old.PluginReport;
 import org.apache.maven.rtinfo.RuntimeInformation;
 import org.apache.maven.tools.plugin.EnhancedParameterWrapper;
 import org.codehaus.plexus.configuration.PlexusConfiguration;
@@ -33,7 +34,7 @@ import org.codehaus.plexus.configuration.PlexusConfigurationException;
 /**
  * Reads enhanced plugin.xml files as generated by
  * {@link org.apache.maven.tools.plugin.generator.PluginDescriptorFilesGenerator} and
- * used by {@link org.apache.maven.plugin.plugin.report.PluginReport}.
+ * used by {@link PluginReport}.
  * Populates the slightly extended {@link Parameter} object {@link EnhancedParameterWrapper}.
  */
 @Deprecated
